+++ app/src/feature/admin/allMembers/index/table/ActiveMemberRow.tsx
@@ -1,8 +1,8 @@
 import type { ActiveMember } from "@/core/domains/member/activeMember";
 import type { FC } from "react";
 import { ListItem } from "../list/ListItem";
-import { PopupMenu } from "../popup/Popup";
-import { ThreeDotsIcon } from "../threeDotsIcon/ThreeDots.icon";
+import { Popup } from "../popup/Popup2";
+import { ThreeDotsIcon } from "../threeDotsIcon/ThreeDotsIcon";
 import type { OnClickBan, OnClickDisable } from "./type";
 
 export const ActiveMemberRow: FC<{
@@ -22,13 +22,13 @@ export const ActiveMemberRow: FC<{
 			<td className="cursor-pointer py-2 px-4 border-b border-gray-300">
 				<div className="flex justify-between">
 					<div className="items-center">
-						<PopupMenu opener={<ThreeDotsIcon />}>
+						<Popup opener={<ThreeDotsIcon />}>
 							<ListItem onClick={() => onClickBan(member)} label="Ban" />
 							<ListItem
 								onClick={() => onClickDisable(member)}
 								label="Disable"
 							/>
-						</PopupMenu>
+						</Popup>
 					</div>
 				</div>
 			</td>
