+++ app/src/feature/admin/allMembers/index/table/PendingActivationMemberRow.tsx
@@ -1,8 +1,8 @@
 import type { PendingActivationMember } from "@/core/domains/member/pendingActivationMember";
 import type { FC } from "react";
 import { ListItem } from "../list/ListItem";
-import { PopupMenu } from "../popup/Popup";
-import { ThreeDotsIcon } from "../threeDotsIcon/ThreeDots.icon";
+import { Popup } from "../popup/Popup2";
+import { ThreeDotsIcon } from "../threeDotsIcon/ThreeDotsIcon";
 import type { OnClickBan } from "./type";
 
 export const PendingActivationMemberRow: FC<{
@@ -21,14 +21,14 @@ export const PendingActivationMemberRow: FC<{
 			<td className="cursor-pointer py-2 px-4 border-b border-gray-300">
 				<div className="flex justify-between">
 					<div className="items-center">
-						<PopupMenu opener={<ThreeDotsIcon />}>
+						<Popup opener={<ThreeDotsIcon />}>
 							<ListItem
 								onClick={() => {
 									onClickBan(member);
 								}}
 								label="Ban"
 							/>
-						</PopupMenu>
+						</Popup>
 					</div>
 				</div>
 			</td>
